chore: remove old tip and old nft
