+++ src/commands/index.ts
@@ -55,8 +55,7 @@ import moniker from "./moniker"
 import gas from "./gas"
 import activity from "./activity"
 import inbox from "./inbox"
-// import nft from "./nft"
-import nft from "./new-nft"
+import nft from "./nft"
 import nftrole from "./nft-role"
 import poe from "./poe"
 import profile from "./profile"
@@ -69,8 +68,7 @@ import starboard from "./starboard"
 import statements from "./statements"
 import telegram from "./telegram"
 import ticker from "./ticker"
-// import tip from "./tip"
-import tip from "./new-tip"
+import tip from "./tip"
 import token from "./token"
 import top from "./top"
 import verify from "./verify"
