--- src/commands/new-nft/list/slash.ts
@@ -1,1 +1,0 @@
-// TODO: slash version
