--- src/commands/new-nft/query/slash.ts
@@ -1,1 +1,0 @@
-// TODO: slash version
