--- src/commands/nft/integrate/processor.test.ts
@@ -1,104 +1,0 @@
-import * as processor from "./processor"
-import { getErrorEmbed, getSuccessEmbed } from "ui/discord/embed"
-import { assertRunResult } from "../../../../tests/assertions/discord"
-import community from "adapters/community"
-jest.mock("adapters/community")
-
-describe("executeNftIntegrateCommand", () => {
-  afterEach(() => jest.clearAllMocks())
-
-  test("executeNftIntegrateCommand successfully", async () => {
-    jest.spyOn(community, "getNftCollectionInfo").mockResolvedValueOnce({
-      status: 200,
-      data: {
-        symbol: "ABC",
-        name: "ABC",
-      },
-    } as any)
-    jest.spyOn(community, "updateSupportVerse").mockResolvedValueOnce({
-      status: 200,
-    } as any)
-    const output = await processor.executeNftIntegrateCommand(
-      "0x123123",
-      "5",
-      "123123",
-      "121212",
-      undefined
-    )
-    const expected = {
-      messageOptions: {
-        embeds: [
-          getSuccessEmbed({
-            title: `ABC integrated`,
-            description: `ABC collection is now ready to take part in our verse (added + enabled)`,
-          }),
-        ],
-      },
-    }
-
-    expect(community.getNftCollectionInfo).toHaveBeenCalled()
-    expect(community.updateSupportVerse).toHaveBeenCalled()
-    assertRunResult(output, expected)
-  })
-
-  test("executeNftIntegrateCommand api 500", async () => {
-    jest.spyOn(community, "getNftCollectionInfo").mockResolvedValueOnce({
-      status: 200,
-      data: {},
-    } as any)
-    jest.spyOn(community, "updateSupportVerse").mockResolvedValueOnce({
-      status: 500,
-    } as any)
-    const output = await processor.executeNftIntegrateCommand(
-      "0x123123",
-      "5",
-      "123123",
-      "121212",
-      undefined
-    )
-    const expected = {
-      messageOptions: {
-        embeds: [
-          getErrorEmbed({
-            description: "Internal Server Error",
-          }),
-        ],
-      },
-    }
-
-    expect(community.getNftCollectionInfo).toHaveBeenCalled()
-    expect(community.updateSupportVerse).toHaveBeenCalled()
-    assertRunResult(output, expected)
-  })
-
-  test("executeNftIntegrateCommand api 400", async () => {
-    jest.spyOn(community, "getNftCollectionInfo").mockResolvedValueOnce({
-      status: 200,
-      data: {},
-    } as any)
-    jest.spyOn(community, "updateSupportVerse").mockResolvedValueOnce({
-      status: 400,
-      error: "Error",
-    } as any)
-    const output = await processor.executeNftIntegrateCommand(
-      "0x123123",
-      "5",
-      "123123",
-      "121212",
-      undefined
-    )
-    const expected = {
-      messageOptions: {
-        embeds: [
-          getErrorEmbed({
-            description: "Error",
-          }),
-        ],
-      },
-    }
-
-    expect(community.getNftCollectionInfo).toHaveBeenCalled()
-    expect(community.updateSupportVerse).toHaveBeenCalled()
-    assertRunResult(output, expected)
-  })
-})
