+++ src/commands/nft/integrate/processor.ts
@@ -1,17 +1,20 @@
 import community from "adapters/community"
-import { Message } from "discord.js"
+import { CommandInteraction, Message } from "discord.js"
 import { InternalError } from "errors"
 import { getErrorEmbed, getSuccessEmbed } from "ui/discord/embed"
 import { getEmoji } from "utils/common"
 import { callAPI, toEmbed } from "../processor"
+import { reply } from "utils/discord"
 
 export async function executeNftIntegrateCommand(
   address: string,
   chainId: string,
   authorId: string,
   guildId: string,
-  msg: Message | undefined
+  msgOrInteraction: Message | CommandInteraction
 ) {
+  let response
+  const msg = msgOrInteraction instanceof Message ? msgOrInteraction : undefined
   // check existed collection address - chainId
   const { status, data } = await community.getNftCollectionInfo({
     address,
@@ -29,13 +32,14 @@ export async function executeNftIntegrateCommand(
     )
     // return early if the `add` command didn't succeed
     if (storeCollectionRes.status !== 200) {
-      return toEmbed(storeCollectionRes, supportedChainsRes, msg)
+      response = await toEmbed(storeCollectionRes, supportedChainsRes, msg)
+      await reply(msgOrInteraction, response)
     }
   }
 
   if (!data) {
     throw new InternalError({
-      msgOrInteraction: msg,
+      msgOrInteraction: msgOrInteraction,
       title: "Can't find the NFT collection",
       description: `The NFT Address and NFT Chain must be valid. Go to the collection's official website/marketplace to find this information.\n${getEmoji(
         "ANIMATED_POINTING_RIGHT",
@@ -47,7 +51,7 @@ export async function executeNftIntegrateCommand(
   const { status: supportVerseStatus = 0, error } =
     await community.updateSupportVerse(address)
   if (supportVerseStatus === 200) {
-    return {
+    response = {
       messageOptions: {
         embeds: [
           getSuccessEmbed({
@@ -59,6 +63,7 @@ export async function executeNftIntegrateCommand(
         ],
       },
     }
+    await reply(msgOrInteraction, response)
   } else {
     let description
     if ([400, 500].includes(supportVerseStatus)) {
@@ -68,7 +73,7 @@ export async function executeNftIntegrateCommand(
         description = error ?? ""
       }
     }
-    return {
+    response = {
       messageOptions: {
         embeds: [
           getErrorEmbed({
@@ -78,5 +83,6 @@ export async function executeNftIntegrateCommand(
         ],
       },
     }
+    await reply(msgOrInteraction, response)
   }
 }
