+++ src/commands/nft/processor.ts
@@ -88,7 +88,6 @@ export async function toEmbed(
 ) {
   // get response and show discord message
   const { error } = storeCollectionRes
-  const { data: chainData } = await supportedChainsRes
   switch (storeCollectionRes.status) {
     case 200:
       return buildDiscordMessage(
@@ -100,82 +99,7 @@ export async function toEmbed(
     case 500:
       return buildDiscordMessage(msg, "NFT", "Internal Server Error")
     default:
-      if (
-        error.includes(
-          "Cannot get name and symbol of contract: This collection does not support collection name"
-        )
-      ) {
-        return buildDiscordMessage(
-          msg,
-          "NFT",
-          "This collection does not support collection name."
-        )
-      } else if (
-        error.includes(
-          "Cannot get name and symbol of contract: This collection does not support collection symbol"
-        )
-      ) {
-        return buildDiscordMessage(
-          msg,
-          "NFT",
-          "This collection does not support collection symbol."
-        )
-      } else if (
-        error.includes(
-          "Cannot get name and symbol of contract: no contract code at given address"
-        )
-      ) {
-        throw new InternalError({
-          msgOrInteraction: msg,
-          title: "Can't find the NFT collection",
-          description:
-            "The NFT Address and NFT Chain must be valid. Go to the collection's official website/ marketplace to find this information. ",
-        })
-      } else if (error.includes("Already added. Nft is in sync progress")) {
-        return buildDiscordMessage(
-          msg,
-          "Existing Collection",
-          "Please add another one or view the collection by `$nft <collection_symbol> <token_id>`."
-        )
-      } else if (error.includes("block number not synced yet")) {
-        return buildDiscordMessage(msg, "NFT", "Block number is not in sync.")
-      } else if (error.includes("Already added. Nft is done with sync")) {
-        return buildDiscordMessage(
-          msg,
-          "NFT",
-          "Already added. Nft is done with sync"
-        )
-      } else if (error.includes("chain is not supported/invalid")) {
-        // add list chain to description
-        const listChainSupportedPrefix = `List chain supported:\n`
-        let listChainSupported = ""
-        for (const chainItm of chainData) {
-          listChainSupported = listChainSupported + `${chainItm}\n`
-        }
-        const listChainDescription =
-          `Chain is not supported. ` +
-          listChainSupportedPrefix +
-          "```\n" +
-          listChainSupported +
-          "```"
-        return buildDiscordMessage(msg, "NFT", listChainDescription)
-      } else if (
-        error.includes("duplicate key value violates unique constraint")
-      ) {
-        return buildDiscordMessage(
-          msg,
-          "NFT",
-          "This collection is already added"
-        )
-      } else if (error.includes("No metadata found")) {
-        return buildDiscordMessage(
-          msg,
-          "NFT",
-          "Cannot found metadata for this collection"
-        )
-      } else {
-        return buildDiscordMessage(msg, "NFT", error)
-      }
+      return buildDiscordMessage(msg, "NFT", error)
   }
 }
 
@@ -184,7 +108,7 @@ export async function composeCollectionInfoEmbed(
   collectionAddress: string,
   chain: string
 ) {
-  if (chain === "999" || chain === "sol") {
+  if (chain == "sol" || chain == "999") {
     collectionAddress = "solscan-" + collectionAddress
   }
   const { data, ok, curl, log } = await community.getNFTCollectionMetadata(
@@ -192,7 +116,7 @@ export async function composeCollectionInfoEmbed(
     chain
   )
   if (!ok) {
-    throw new APIError({ msgOrInteraction: msg, curl: curl, description: log })
+    throw new APIError({ msgOrInteraction: msg, curl, description: log })
   }
   if (!data) {
     throw new InternalError({
