+++ src/commands/tip/index/parser.test.ts
