+++ src/commands/tip/index/processor.ts
