+++ src/commands/tip/index/slash.test.ts
