+++ src/commands/tip/index/slash.ts
