+++ src/commands/tip/index/text.test.ts
