+++ src/commands/tip/index/text.ts
