+++ src/commands/tip/telegram/processor.ts
