+++ src/commands/tip/telegram/text.ts
